+++ lib/conjur/escape.rb
@@ -80,9 +80,8 @@ module Conjur
         return "false" unless str
         str = str.id if str.respond_to?(:id)
         # Leave colons and forward slashes alone
-        require 'uri'
-        pattern = URI::PATTERN::UNRESERVED + ":\\/@"
-        URI.escape(str.to_s, Regexp.new("[^#{pattern}]"))
+        require 'addressable/uri'
+        Addressable::URI.encode(str.to_s)
       end
     end